Separated double U-joint into a different file.
